Backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
gdown
pydantic